Releasing 0.8.0 (#356)
* fix(ci): pin to specific stable rust version (#287)

This commit pins the rust-toolchain to a specific rust version (1.89.0)
for compatibility guarantees. Renovate will handle warning on outdated
versions of stable rust.

A minimal supported rust version was also added to the underlying crates
to ensure compatibility guarantees as the stable verison increments in
future updates.

* feat: Implement Test Coverage Measurement and Reporting

* chore: add changeset

* feat: remove unnecessary codecov setting

* feat: add codecov badge

* feat: add codecov config file

* feat: add code coverage docs to CONTRIBUTING.md

* test: add tests for uplink schema event

* ci: prevent draft PRs from verifying changeset

* chore: add changeset

* ci: trigger verify changeset workflow when PR becomes ready for review

* ci: update github checkout actions

* feat: Configuration for disabling authorization token passthrough (#336)

* Allow config for not forwarding Auth tokens to GraphQL API

* use serde default to make new config optional

* Changeset

* Fix default case in main.rs

* Add new config option to documentation

* Redirect /docs/apollo-mcp-server/guides to fix 404 (#349)


* chore: update RMCP version to latest (#328)

This commit updates the RMCP dependency to the latest version 0.6.4.
Sadly, schemars was also updated with this, so a lot of unrelated
changes were needed to conform with the new stable schemars version.

* chore(release): bumping to version 0.8.0

* chore(release): changelog for 0.8.0

---------
