+++ .github/workflows/verify-changeset.yml
@@ -1,6 +1,7 @@
 name: Verify Changeset
 on:
   pull_request:
+    types: [opened, reopened, synchronize, ready_for_review]
     branches-ignore:
       - main
       - release/**
@@ -19,7 +20,7 @@ on:
 
 jobs:
   verify-changeset:
-    if: ${{ !contains(github.event.pull_request.labels.*.name, 'skip-changeset') && !startsWith(github.head_ref, 'sync/') && !startsWith(github.head_ref, 'conflict/') }}
+    if: ${{ !contains(github.event.pull_request.labels.*.name, 'skip-changeset') && !startsWith(github.head_ref, 'sync/') && !startsWith(github.head_ref, 'conflict/') && !github.event.pull_request.draft }}
     name: Verify
     runs-on: ubuntu-24.04
     permissions:
