+++ Cargo.toml
@@ -8,7 +8,11 @@ members = [
 
 [workspace.package]
 authors = ["Apollo <opensource@apollographql.com>"]
-version = "0.7.5"
+edition = "2024"
+license-file = "LICENSE"
+repository = "https://github.com/apollographql/apollo-mcp-server"
+rust-version = "1.89.0"
+version = "0.8.0"
 
 [workspace.dependencies]
 apollo-compiler = "1.27.0"
