+++ crates/apollo-mcp-server/src/auth.rs
@@ -46,6 +46,10 @@ pub struct Config {
 
     /// Supported OAuth scopes by this resource server
     pub scopes: Vec<String>,
+
+    /// Whether to disable the auth token passthrough to upstream API
+    #[serde(default)]
+    pub disable_auth_token_passthrough: bool,
 }
 
 impl Config {
