+++ crates/apollo-mcp-server/src/main.rs
@@ -109,6 +109,8 @@ async fn main() -> anyhow::Result<()> {
         .then(|| config.graphos.graph_ref())
         .transpose()?;
 
+    let transport = config.transport.clone();
+
     Ok(Server::builder()
         .transport(config.transport)
         .schema_source(schema_source)
@@ -125,6 +127,15 @@ async fn main() -> anyhow::Result<()> {
         .mutation_mode(config.overrides.mutation_mode)
         .disable_type_description(config.overrides.disable_type_description)
         .disable_schema_description(config.overrides.disable_schema_description)
+        .disable_auth_token_passthrough(match transport {
+            apollo_mcp_server::server::Transport::Stdio => false,
+            apollo_mcp_server::server::Transport::SSE { auth, .. } => auth
+                .map(|a| a.disable_auth_token_passthrough)
+                .unwrap_or(false),
+            apollo_mcp_server::server::Transport::StreamableHttp { auth, .. } => auth
+                .map(|a| a.disable_auth_token_passthrough)
+                .unwrap_or(false),
+        })
         .custom_scalar_map(
             config
                 .custom_scalars
