+++ crates/apollo-mcp-server/src/server.rs
@@ -36,6 +36,7 @@ pub struct Server {
     mutation_mode: MutationMode,
     disable_type_description: bool,
     disable_schema_description: bool,
+    disable_auth_token_passthrough: bool,
     search_leaf_depth: usize,
     index_memory_bytes: usize,
     health_check: HealthCheckConfig,
@@ -112,6 +113,7 @@ impl Server {
         mutation_mode: MutationMode,
         disable_type_description: bool,
         disable_schema_description: bool,
+        disable_auth_token_passthrough: bool,
         search_leaf_depth: usize,
         index_memory_bytes: usize,
         health_check: HealthCheckConfig,
@@ -138,6 +140,7 @@ impl Server {
             mutation_mode,
             disable_type_description,
             disable_schema_description,
+            disable_auth_token_passthrough,
             search_leaf_depth,
             index_memory_bytes,
             health_check,
