+++ crates/apollo-mcp-server/src/server/states/running.rs
@@ -52,6 +52,7 @@ pub(super) struct Running {
     pub(super) mutation_mode: MutationMode,
     pub(super) disable_type_description: bool,
     pub(super) disable_schema_description: bool,
+    pub(super) disable_auth_token_passthrough: bool,
     pub(super) health_check: Option<HealthCheck>,
 }
 
@@ -211,7 +212,9 @@ impl ServerHandler for Running {
                 let mut headers = self.headers.clone();
                 if let Some(axum_parts) = context.extensions.get::<axum::http::request::Parts>() {
                     // Optionally extract the validated token and propagate it to upstream servers if present
-                    if let Some(token) = axum_parts.extensions.get::<ValidToken>() {
+                    if !self.disable_auth_token_passthrough
+                        && let Some(token) = axum_parts.extensions.get::<ValidToken>()
+                    {
                         headers.typed_insert(token.deref().clone());
                     }
 
@@ -242,7 +245,9 @@ impl ServerHandler for Running {
                 let mut headers = self.headers.clone();
                 if let Some(axum_parts) = context.extensions.get::<axum::http::request::Parts>() {
                     // Optionally extract the validated token and propagate it to upstream servers if present
-                    if let Some(token) = axum_parts.extensions.get::<ValidToken>() {
+                    if !self.disable_auth_token_passthrough
+                        && let Some(token) = axum_parts.extensions.get::<ValidToken>()
+                    {
                         headers.typed_insert(token.deref().clone());
                     }
 
@@ -302,7 +307,12 @@ impl ServerHandler for Running {
         ServerInfo {
             server_info: Implementation {
                 name: "Apollo MCP Server".to_string(),
+                icons: None,
+                title: Some("Apollo MCP Server".to_string()),
                 version: env!("CARGO_PKG_VERSION").to_string(),
+                website_url: Some(
+                    "https://www.apollographql.com/docs/apollo-mcp-server".to_string(),
+                ),
             },
             capabilities: ServerCapabilities::builder()
                 .enable_tools()
@@ -355,6 +365,7 @@ mod tests {
             mutation_mode: MutationMode::None,
             disable_type_description: false,
             disable_schema_description: false,
+            disable_auth_token_passthrough: false,
             health_check: None,
         };
