+++ docs/source/install.mdx
@@ -26,14 +26,14 @@ To download a **specific version** of Apollo MCP Server (recommended for CI envi
 
 ```bash
 # Note the `v` prefixing the version number
-docker image pull ghcr.io/apollographql/apollo-mcp-server:v0.7.5
+docker image pull ghcr.io/apollographql/apollo-mcp-server:v0.8.0
 ```
 
 To download a specific version of Apollo MCP Server that is a release candidate:
 
 ```bash
 # Note the `v` prefixing the version number and the `-rc` suffix
-docker image pull ghcr.io/apollographql/apollo-mcp-server:v0.7.5-rc.1
+docker image pull ghcr.io/apollographql/apollo-mcp-server:v0.8.0-rc.1
 ```
 
 <Note>
@@ -65,7 +65,7 @@ To install or upgrade to a **specific version** of Apollo MCP Server (recommende
 
 ```bash
 # Note the `v` prefixing the version number
-curl -sSL https://mcp.apollo.dev/download/nix/v0.7.5 | sh
+curl -sSL https://mcp.apollo.dev/download/nix/v0.8.0 | sh
 ```
 
 If your machine doesn't have the `curl` command, you can get the latest version from the [`curl` downloads page](https://curl.se/download.html).
@@ -82,5 +82,5 @@ To install or upgrade to a **specific version** of Apollo MCP Server (recommende
 
 ```bash
 # Note the `v` prefixing the version number
-iwr 'https://mcp.apollo.dev/download/win/v0.7.5' | iex
+iwr 'https://mcp.apollo.dev/download/win/v0.8.0' | iex
 ```
