+++ rust-toolchain.toml
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "stable"
+channel = "1.89.0"
 profile = "default"
 components = ["rust-analyzer", "rust-src"]
